refactor: s/Dashboard/Search/g
